add new smart locators and smart templates + create default methods for all customizable actions
