+++ jdi-light-applitools-tests/pom.xml
@@ -31,7 +31,7 @@
         <dependency>
             <groupId>com.epam.jdi</groupId>
             <artifactId>jdi-light-html</artifactId>
-            <version>1.1.15</version>
+            <version>1.1.16</version>
         </dependency>
         <dependency>
             <groupId>com.epam.jdi</groupId>
