+++ jdi-light-eyes/pom.xml
@@ -19,7 +19,7 @@
         <dependency>
             <groupId>com.epam.jdi</groupId>
             <artifactId>jdi-light</artifactId>
-            <version>1.1.15</version>
+            <version>1.1.16</version>
         </dependency>
 
         <!-- Applitools SDK -->
