+++ jdi-light-html-tests/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <version>1.1.15</version>
+    <version>1.1.16</version>
     <artifactId>jdi-light-html-tests</artifactId>
     <name>JDI Html Tests</name>
     <groupId>com.epam.jdi</groupId>
@@ -23,7 +23,7 @@
         <dependency>
             <groupId>com.epam.jdi</groupId>
             <artifactId>jdi-light-html</artifactId>
-            <version>1.1.15</version>
+            <version>1.1.16</version>
         </dependency>
 
         <!--Allure config in parent-->
