+++ jdi-light-html/pom.xml
@@ -7,7 +7,7 @@
     <artifactId>jdi-light-html</artifactId>
     <name>JDI Light Html Elements</name>
     <packaging>jar</packaging>
-    <version>1.1.15</version>
+    <version>1.1.16</version>
     <groupId>com.epam.jdi</groupId>
 
     <properties>
@@ -20,7 +20,7 @@
         <dependency>
             <groupId>com.epam.jdi</groupId>
             <artifactId>jdi-light</artifactId>
-            <version>1.1.15</version>
+            <version>1.1.16</version>
         </dependency>
     </dependencies>
