+++ jdi-light/pom.xml
@@ -7,7 +7,7 @@
     <artifactId>jdi-light</artifactId>
     <name>JDI Light</name>
     <packaging>jar</packaging>
-    <version>1.1.15</version>
+    <version>1.1.16</version>
     <groupId>com.epam.jdi</groupId>
 
     <properties>
