+++ jdi-light/src/main/java/com/epam/jdi/light/elements/pageobjects/annotations/locators/UI.java
@@ -10,7 +10,7 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.FIELD})
 public @interface UI {
-    String value();
+    String value() default "";
     String group() default "";
 
     @Retention(RetentionPolicy.RUNTIME)
