+++ jdi-light/src/main/java/com/epam/jdi/light/elements/pageobjects/annotations/smart/SClass.java
@@ -1,0 +1,15 @@
+package com.epam.jdi.light.elements.pageobjects.annotations.smart;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Created by Roman Iovlev on 14.02.2018
+ * Email: roman.iovlev.jdi@gmail.com; Skype: roman.iovlev
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.FIELD})
+public @interface SClass {
+}
