+++ jdi-performance/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.epam.jdi</groupId>
     <artifactId>jdi-performance</artifactId>
-    <version>1.1.15</version>
+    <version>1.1.16</version>
     <name>JDI Performance</name>
 
     <properties>
@@ -25,7 +25,7 @@
         <dependency>
             <groupId>com.epam.jdi</groupId>
             <artifactId>jdi-light-html</artifactId>
-            <version>1.1.15</version>
+            <version>1.1.16</version>
         </dependency>
 
         <!--Allure-->
